refactor(agents): relocate emoji list to shared; reclassify collaborators as services
- Move list-agent-emoji out of channels/ into shared/emoji (it is reaction
  reference-data served by GET /agents/emoji, not channel onboarding)
- Reclassify CleanupNovuEmail, FindOrCreateNovuEmail, and
  GetMcpNovuAppCredentials as services (NovuEmailCleanupService,
  NovuEmailProvisioningService, McpNovuAppCredentialsService). They are
  collaborators invoked by other usecases, not request handlers with
  commands, so they no longer masquerade as CQRS usecases in USE_CASES.
