+++ apps/api/src/app/agents/shared/emoji/list-agent-emoji/list-agent-emoji.usecase.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@nestjs/common';
 import type { EmojiFormats } from 'chat';
-import { esmImport } from '../../../shared/util/esm-import';
+import { esmImport } from '../../util/esm-import';
 
 export interface AgentEmojiEntry {
   name: string;
